+++ ortools/sat/samples/SolveAndPrintIntermediateSolutionsSampleSat.java
@@ -71,7 +71,7 @@ public class SolveAndPrintIntermediateSolutionsSampleSat {
 
     // Maximize a linear combination of variables.
     // [START objective]
-    model.maximize(LinearExpr.scalProd(new IntVar[] {x, y, z}, new int[] {1, 2, 3}));
+    model.maximize(LinearExpr.scalProd(new IntVar[] {x, y, z}, new long[] {1, 2, 3}));
     // [END objective]
 
     // Create a solver and solve the model.
